parser: Move DefaultFuncs to parser/defaults Also moved Function to parser/values to prevent cyclic imports.
