+++ parser/defaults/functions.go
@@ -1,39 +1,10 @@
-package parser
+package defaults
 
 import (
-	"github.com/syzkrash/skol/parser/nodes"
 	"github.com/syzkrash/skol/parser/values"
 )
 
-type Function struct {
-	Name string
-	Args []values.FuncArg
-	Ret  *values.Type
-}
-
-func DefinedFunction(n *nodes.FuncDefNode) *Function {
-	return &Function{
-		Name: n.Name,
-		Args: n.Args,
-		Ret:  n.Ret,
-	}
-}
-
-func ExternFunction(n *nodes.FuncExternNode) *Function {
-	var name string
-	if n.Intern != "" {
-		name = n.Intern
-	} else {
-		name = n.Name
-	}
-	return &Function{
-		Name: name,
-		Args: n.Args,
-		Ret:  n.Ret,
-	}
-}
-
-var DefaultFuncs = map[string]*Function{
+var Functions = map[string]*values.Function{
 	"print": {
 		Name: "print",
 		Args: []values.FuncArg{{"a", values.String}},
